+++ src/pages/CoffeeChatListPage/components/AddCoffeeChatModal/AddCoffeeChatModal.tsx
@@ -2,7 +2,7 @@ import { CounterMinusIcon, CounterPlusIcon } from "@/assets/svg";
 import { Button, FileUploadBox, HashtagInput, Input, SideModal, TagChip, Textarea } from "@/components";
 import { useFileUpload } from "@/components/FileUploadBox/hooks/useFileUpload";
 
-import { useHashtag } from "@/components/HashtagChip/hooks/useHashtag";
+import { useHashtag } from "@/components/TagChip/hooks/useHashtag";
 import type { SideModalProps } from "@/components/SideModal/types/sideModalTypes";
 import { SUPPORTING_TEXT } from "@/constants/supportingText";
 import { TimeDropdown } from "@/pages/CoffeeChatListPage/components/TimeDropdown/TimeDropdown";
